content/media/webaudio/test/test_AudioContext.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test whether we can create an AudioContext interface</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();
addLoadEvent(function() {
  var ac = new mozAudioContext();
  ok(ac, "Create a mozAudioContext object");
  SimpleTest.finish();
});

</script>
</pre>
</body>
</html>
